Worked on HoldableItemNetwork
