+++ Nightmare Library/Assets/Jack/_Scripts/Interactables/HoldableItemNetwork.cs
@@ -28,9 +28,9 @@ public class HoldableItemNetwork : NetworkBehaviour
     private bool wasUpdating = false;
     private int currentUpdateFrame = 0;
 
-    private NetworkVariable<TransformDataRB> transformData = new NetworkVariable<TransformDataRB>();
+    private NetworkVariable<TransformData> transformData = new NetworkVariable<TransformData>();
     private NetworkVariable<bool> isActive = new NetworkVariable<bool>();
-    private NetworkVariable<bool> isHeld = new NetworkVariable<bool>();
+    private NetworkVariable<HeldData> isHeld = new NetworkVariable<HeldData>();
 
     protected virtual void Awake()
     {
@@ -45,7 +45,6 @@ public class HoldableItemNetwork : NetworkBehaviour
             previousPosition = transform.position;
         }
     }
-
     public override void OnNetworkSpawn()
     {
         base.OnNetworkSpawn();
@@ -63,18 +62,14 @@ public class HoldableItemNetwork : NetworkBehaviour
 
         if (!IsOwner)
         {
-            transformData.OnValueChanged += ConsumeTransformData;
-            isActive.OnValueChanged += ConsumeEnabledData;
+            isActive.OnValueChanged += OnActiveValueChanged;
             isHeld.OnValueChanged += OnHeldValueChanged;
-
-            // Moves the item to the correct location according to the server
-            ConsumeTransformData(transformData.Value, transformData.Value);
         }
         else
         {
             TransmitTransformData();
             isActive.Value = gameObject.activeInHierarchy;
-            isHeld.Value = false;
+            isHeld.Value = new HeldData(false, 0);
         }
     }
 
@@ -83,36 +78,26 @@ public class HoldableItemNetwork : NetworkBehaviour
         // Check to make sure the network is running to avoid calls going out without being connected and that this is on the server/owner
         if (NetworkConnectionController.IsRunning)
         {
-            if (IsOwner && canUpdateRigidbody && isActive.Value)
+            // Check to make sure that this object has a rigid body, is active and is being run on the server
+            if (NetworkManager.IsServer && canUpdateRigidbody && isActive.Value)
             {
-                // ensures the update only runs every few frames
-                if (currentUpdateFrame >= updateTransformFrequency)
+                // Check if the object is moving enough to pass the velocity check
+                if (Vector3.SqrMagnitude(previousPosition - transform.position) > movementThreshold * movementThreshold)
                 {
-                    // Check if the object is moving enough to pass the velocity check
-                    if (Vector3.SqrMagnitude(previousPosition - transform.position) > movementThreshold * movementThreshold)
-                    {
-                        // Transmit this object's transform data
-                        TransmitTransformData();
-                        wasUpdating = true;
-                    }
-                    // Sends one more update just as the object stops moving
-                    else if (wasUpdating)
-                    {
-                        // Transmit this object's transform data
-                        TransmitTransformData();
-                        RectifyTransform();
-                        wasUpdating = false;
-                    }
+                    // Transmit this object's transform data
+                    TransmitTransformData();
+                    wasUpdating = true;
+                }
+                // Sends one more update just as the object stops moving
+                else if (wasUpdating)
+                {
+                    // Transmit this object's transform data
+                    TransmitTransformData();
+                    RectifyTransform();
+                    wasUpdating = false;
+                }
 
-                    // Reset the frame counter
-                    currentUpdateFrame = 0;
-                    previousPosition = transform.position;
-                }
-                else
-                {
-                    // Increment the frame counter
-                    currentUpdateFrame++;
-                }
+                previousPosition = transform.position;
             }
             else if (!IsOwner && parent.hasRigidBody && parent.rb.velocity == Vector3.zero)
             {
@@ -127,46 +112,40 @@ public class HoldableItemNetwork : NetworkBehaviour
         }
     }
 
-    private void ConsumeEnabledData(bool previousValue, bool newValue)
-    {
-        parent.gameObject.SetActive(newValue);
-    }
-
     #region Transform
 
     private void TransmitTransformData()
     {
-        var state = new TransformDataRB
+        var state = new TransformData
         {
             Position = parent.trans.position,
             Rotation = parent.trans.rotation,
             Velocity = parent.hasRigidBody ? parent.rb.velocity : Vector3.zero
         };
 
-        // Needed because we are not able to change info if server has authority
+        // just a safety net in case the client somehow is able to call this method
         if (IsServer)
         {
             transformData.Value = state;
-        }
-        else
-        {
-            TransmitTransformDataServerRpc(state);
+            TransformDataUpdateClientRpc();
         }
     }
-    [ServerRpc(RequireOwnership = false)]
-    private void TransmitTransformDataServerRpc(TransformDataRB state)
+    [ClientRpc]
+    private void TransformDataUpdateClientRpc()
     {
-        transformData.Value = state;
+        // May need to pass data through method as well
+        ConsumeTransformData();
     }
-    private void ConsumeTransformData(TransformDataRB previousValue, TransformDataRB newValue)
+
+    private void ConsumeTransformData()
     {
         // Ensure that the owner does not waste time updating to it's own values
         if (!IsOwner)
         {
-            if (Vector3.SqrMagnitude(newValue.Position - transform.position) < transformThreshold * transformThreshold)
+            if (Vector3.SqrMagnitude(transformData.Value.Position - transform.position) < transformThreshold * transformThreshold)
             {
-                transform.rotation = Quaternion.Lerp(parent.trans.rotation, newValue.Rotation, interpolationStrength);
-                parent.trans.position = Vector3.Slerp(parent.trans.position, newValue.Position, interpolationStrength);
+                transform.rotation = Quaternion.Lerp(parent.trans.rotation, transformData.Value.Rotation, interpolationStrength);
+                parent.trans.position = Vector3.Slerp(parent.trans.position, transformData.Value.Position, interpolationStrength);
                 if (parent.hasRigidBody && !parent.rb.isKinematic)
                 {
                     parent.rb.velocity = transformData.Value.Velocity;
@@ -240,11 +219,10 @@ public class HoldableItemNetwork : NetworkBehaviour
             if (IsOwner)
             {
                 TransmitTransformData();
-                isHeld.Value = true;
-                PlaceClientRpc(new TransformData(parent.trans.position, parent.trans.rotation), NetworkManager.LocalClientId);
+                PlaceClientRpc(new TransformData(parent.trans.position, parent.trans.rotation, Vector3.zero), NetworkManager.LocalClientId);
             }
             else
-                TransmitPlaceServerRpc(NetworkManager.LocalClientId, new TransformData(parent.trans.position, parent.trans.rotation));
+                TransmitPlaceServerRpc(NetworkManager.LocalClientId, new TransformData(parent.trans.position, parent.trans.rotation, Vector3.zero));
         }
     }
     [ServerRpc(RequireOwnership = false)]
@@ -255,7 +233,6 @@ public class HoldableItemNetwork : NetworkBehaviour
         PlaceClientRpc(data, sender);
 
         TransmitTransformData();
-        isHeld.Value = true;
     }
     [ClientRpc]
     protected virtual void PlaceClientRpc(TransformData data, ulong sender)
@@ -299,6 +276,12 @@ public class HoldableItemNetwork : NetworkBehaviour
 
     #endregion
 
+    #region Holding Update
+
+
+
+    #endregion
+
     #region Active / Colliders
 
     private void OnActiveChanged(bool b)
@@ -324,22 +307,112 @@ public class HoldableItemNetwork : NetworkBehaviour
     private void OnHeldChanged(bool b)
     {
         if (!NetworkManager.IsServer)
-            OnHeldChangedServerRpc(b);
+            OnHeldChangedServerRpc(new HeldData(b, NetworkManager.LocalClientId));
         else
         {
-            isHeld.Value = b;
-            OnHeldValueChanged(b, b);
+            isHeld.Value = new HeldData(b, NetworkManager.LocalClientId);
+            OnHeldValueChanged(isHeld.Value, isHeld.Value);
         }
             
     }
     [ServerRpc(RequireOwnership = false)]
-    private void OnHeldChangedServerRpc(bool b)
+    private void OnHeldChangedServerRpc(HeldData data)
     {
-        OnHeldChanged(b);
+        isHeld.Value = data;
+        OnHeldValueChanged(data, data);
     }
-    private void OnHeldValueChanged(bool previous, bool current)
+    private void OnHeldValueChanged(HeldData previous, HeldData current)
     {
-        parent.SetHeld(current, true);
+        parent.SetHeld(current.isHeld, true);
+    }
+
+    #endregion
+
+    #region Data
+
+    public struct TransformData : INetworkSerializable
+    {
+        private float xPos, yPos, zPos;
+        private float xRot, yRot, zRot;
+        private float xVel, yVel, zVel;
+
+        internal Vector3 Position
+        {
+            get => new Vector3(xPos, yPos, zPos);
+            set
+            {
+                xPos = value.x;
+                yPos = value.y;
+                zPos = value.z;
+            }
+        }
+        internal Quaternion Rotation
+        {
+            get => Quaternion.Euler(xRot, yRot, zRot);
+            set
+            {
+                xRot = value.eulerAngles.x;
+                yRot = value.eulerAngles.y;
+                zRot = value.eulerAngles.z;
+            }
+        }
+        internal Vector3 Velocity
+        {
+            get => new Vector3(xVel, yVel, zVel);
+            set
+            {
+                xVel = value.x;
+                yVel = value.y;
+                zVel = value.z;
+            }
+        }
+
+        public TransformData(Vector3 pos, Quaternion rot, Vector3 vel)
+        {
+            xPos = pos.x;
+            yPos = pos.y;
+            zPos = pos.z;
+
+            xRot = rot.x;
+            yRot = rot.y;
+            zRot = rot.z;
+
+            xVel = vel.x;
+            yVel = vel.y;
+            zVel = vel.z;
+        }
+
+        public void NetworkSerialize<T>(BufferSerializer<T> serializer) where T : IReaderWriter
+        {
+            serializer.SerializeValue(ref xPos);
+            serializer.SerializeValue(ref yPos);
+            serializer.SerializeValue(ref zPos);
+
+            serializer.SerializeValue(ref xRot);
+            serializer.SerializeValue(ref yRot);
+            serializer.SerializeValue(ref zRot);
+
+            serializer.SerializeValue(ref xVel);
+            serializer.SerializeValue(ref yVel);
+            serializer.SerializeValue(ref zVel);
+        }
+    }
+    public struct HeldData : INetworkSerializable
+    {
+        public bool isHeld;
+        public ulong holderID;
+
+        public HeldData(bool isHeld, ulong holderID)
+        {
+            this.isHeld = isHeld;
+            this.holderID = holderID;
+        }
+
+        public void NetworkSerialize<T>(BufferSerializer<T> serializer) where T : IReaderWriter
+        {
+            serializer.SerializeValue(ref isHeld);
+            serializer.SerializeValue(ref holderID);
+        }
     }
 
     #endregion
