--- Nightmare Library/Assets/Jack/_Scripts/Inventory Scripts/PlayerHoldItemNetwork.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6edc1f10069472347815946127b2d382
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
